Redefined indexes specific to account_operations::op_type_id column..

+++ src/sql_serializer/include/hive/plugins/sql_serializer/accounts_collector.h
@@ -59,18 +59,18 @@ namespace hive::plugins::sql_serializer {
 
       void on_new_account(const hive::protocol::account_name_type& account_name);
 
-      void on_new_operation(const hive::protocol::account_name_type& account_name, int64_t operation_id, int64_t operation_type_id);
+      void on_new_operation(const hive::protocol::account_name_type& account_name, int64_t operation_id, int32_t operation_type_id);
 
     private:
       hive::chain::database& _chain_db;
       cached_data_t& _cached_data;
 
       int64_t _processed_operation_id = -1;
-      int64_t _processed_operation_type_id = -1;
+      int32_t _processed_operation_type_id = -1;
 
       uint32_t _block_num = 0;
 
-      int64_t _creation_operation_type_id = -1;
+      int32_t _creation_operation_type_id = -1;
       fc::optional<int64_t> _creation_operation_id;
 
       flat_set<hive::protocol::account_name_type> _impacted;
